dataframe tests 37.5% broken up -- all run and pass

+++ src/dataframe/column.h
@@ -28,7 +28,7 @@ class Column : public Object {
     * undefined behavior. **/
   virtual void push_back(int val) { return; }
   virtual void push_back(bool val) { return; }
-  virtual void push_back(float val) { return; }
+  virtual void push_back(double val) { return; }
   virtual void push_back(String* val) { return; }
  
  /** Returns the number of elements in the column. */
